requirements.txt
absl-py==0.14.1
aiofiles==0.7.0
aiohttp==3.5.4
APScheduler==3.6.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
azure-core==1.21.1
azure-storage-blob==12.9.0
backports.zoneinfo==0.2.1
boto3==1.9.146
botocore==1.12.253
cached-property==1.5.2
cachetools==4.2.4
certifi==2019.6.16
cffi==1.14.6
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.14.0
constantly==15.1.0
cryptography==35.0.0
cycler==0.10.0
cymem==2.0.5
cytoolz==0.9.0.1
decorator==5.1.0
dill==0.2.9
docopt==0.6.2
docutils==0.15.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.0.0/en_core_web_md-2.0.0.tar.gz
fakeredis==1.0.3
fbmessenger==6.0.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.5.2
gevent==1.4.0
google-auth==2.2.1
google-auth-oauthlib==0.4.6
greenlet==0.4.15
grpcio==1.41.0
gspread==4.0.1
h5py==2.8.0
httplib2==0.19.1
httptools==0.3.0
humanfriendly==4.18
hyperlink==21.0.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
incremental==21.3.0
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
jsonpickle==1.1
jsonschema==2.6.0
kafka-python==1.4.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
klein==17.10.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.0.0
mattermostwrapper==2.1
mock==4.0.3
monotonic==1.5
msgpack==0.5.6
msgpack-numpy==0.4.3.2
msrest==0.6.21
multidict==4.7.6
murmurhash==1.0.5
mysql-connector==2.2.9
networkx==2.3
num2words==0.5.10
numpy==1.16.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==19.0
pandas==1.1.5
pika==1.0.1
plac==0.9.6
preshed==2.0.1
prompt-toolkit==2.0.10
protobuf==3.18.1
psycopg2-binary==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pydub==0.25.1
PyJWT==2.1.0
pykwalify==1.7.0
pyloudnorm==0.1.0
pymongo==3.8.0
pyparsing==2.4.7
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.0
python-engineio==4.2.1
python-socketio==4.0.1
python-telegram-bot==11.1.0
pytz==2019.1
PyYAML==5.4.1
questionary==1.1.1
-e git+https://github.com/Saarthi-ai/pitstop_revamp.git@03b443bee9e3b6171fc60ee04d598777e34350fa#egg=rasa&subdirectory=core
rasa-nlu==0.15.1
-e git+https://github.com/Saarthi-ai/pitstop_revamp.git@03b443bee9e3b6171fc60ee04d598777e34350fa#egg=rasa_sdk&subdirectory=core/rasa_sdk
redis==3.2.1
regex==2018.1.10
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.31
rsa==4.7.2
ruamel.yaml==0.15.94
s3transfer==0.2.1
sanic==19.3.1
Sanic-Cors==0.9.8
sanic-jwt==1.3.1
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.20.2
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.1
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
spacy==2.0.18
SQLAlchemy==1.3.24
tabulate==0.8.9
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==6.12.1
toolz==0.11.1
tqdm==4.31.0
twilio==6.26.3
Twisted==21.7.0
typing==3.7.4
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.2.0
urllib3==1.25.3
uvloop==0.14.0
wcwidth==0.2.5
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==6.0
Werkzeug==0.15.5
wrapt==1.10.11
yarl==1.6.3
zipp==3.6.0
zope.interface==5.4.0
